Break the startSystemSe overload deadlock (from Gemini), five binaries verified
Both startSystemSe overloads exist in the retail binary, but declaring
both made every existing call site ambiguous and failed the build. I
backed it out and recorded it as a dead end. Gemini found the real fix,
and it was not in the header at all.

It audited all seven call sites against the binary rather than reasoning
about them: every one targets FUiUl at 0x801954C0, none targets FUlUl.
So the current default resolution was correct -- but it was correct by
accident, and nobody had checked.

The root cause is that the arguments are int-typed. An int converts to
`unsigned int` and to `unsigned long` at IDENTICAL overload-resolution
rank, so with both overloads visible nothing can resolve. Fixing the
arguments rather than the header:

- Three enum call sites take an explicit (u32) cast.
- d_yes_no_window.cpp's `const int SoundEffects[]` becomes `const u32`,
  which resolves its four call sites with no cast at all and emits the
  same 16 bytes of .rodata.

That is four already-matching TUs edited, which is why it was tested
before it was believed: rebuilt, all five binaries byte-identical, so
zero emitted bytes changed. Gemini predicted exactly that and had
verified it by compiling each TU and diffing the instruction streams
first.

The FUlUl overload is now declared and pinned in syms.txt at 0x801954B0,
which d_a_player_manager.cpp's setHipAttackQuake needs. Note the build
passed BEFORE the pin was added, because a declared-but-uncalled function
needs no definition -- the pin only starts mattering when the new unit
lands and actually calls it.

+++ source/d_profileNP/bases/d_controller_information.cpp
@@ -84,7 +84,7 @@ int dControllerInformation_c::execute() {
                 break;
 
             case END:
-                SndAudioMgr::sInstance->startSystemSe(SE_SYS_BUTTON_SKIP, 1);
+                SndAudioMgr::sInstance->startSystemSe((u32)SE_SYS_BUTTON_SKIP, 1);
                 mLayout.AnimeEndSetup(IN2BTN);
                 mLayout.AnimeEndSetup(LOOP2BTN);
                 mLayout.AnimeStartSetup(HIT2BTN, false);
